Removed class from niftiSave to make code easier.
